HOTFIX & REFACTOR: Refactor of weather pleasantry into mutliple files. HOTFIX for weather order in results table, saturday is a special case because of wraparound
